Split ComponentSpecificContent into smaller parts
