+++ frontend/packages/ux-editor/src/components/config/componentSpecificContent/Address/AddressComponent.module.css
@@ -1,0 +1,5 @@
+.root > div {
+  display: flex;
+  flex-direction: column;
+  gap: var(--fieldset-gap);
+}
