+++ frontend/packages/ux-editor/src/components/config/componentSpecificContent/Address/index.ts
@@ -1,0 +1,1 @@
+export { AddressComponent } from './AddressComponent';
