+++ frontend/packages/ux-editor/src/components/config/componentSpecificContent/Button/index.ts
@@ -1,0 +1,1 @@
+export { ButtonComponent } from './ButtonComponent';
