+++ frontend/packages/ux-editor/src/components/config/componentSpecificContent/FileUpload/FileUploadComponent.module.css
@@ -1,0 +1,5 @@
+.fieldset > div {
+  display: flex;
+  flex-direction: column;
+  gap: var(--fieldset-gap);
+}
