+++ frontend/packages/ux-editor/src/components/config/componentSpecificContent/FileUpload/index.ts
@@ -1,0 +1,1 @@
+export { FileUploadComponent } from './FileUploadComponent';
