--- frontend/packages/ux-editor/src/components/config/componentSpecificContent/FileUploadWithTag/index.ts
@@ -1,1 +1,0 @@
-export { FileUploadWithTagComponent } from './FileUploadWithTagComponent';
