+++ frontend/packages/ux-editor/src/components/config/componentSpecificContent/Image/ImageComponent.module.css
@@ -1,3 +1,9 @@
+.root > div {
+  display: flex;
+  flex-direction: column;
+  gap: var(--fieldset-gap);
+}
+
 .widthAndPlacement {
   display: flex;
   flex-direction: row;
