+++ frontend/packages/ux-editor/src/components/config/editModal/EditDescription.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import type { IGenericEditComponent } from '../componentConfig';
 import { TextResource } from '../../TextResource';
-import { getLanguageFromKey } from 'app-shared/utils/language';
+import { useText } from '../../../hooks';
 
 export const EditDescription = ({
   component,
   handleComponentChange,
-  language,
 }: IGenericEditComponent) => {
+  const t = useText();
   const handleIdChange = (id: string) => handleComponentChange({
     ...component,
     textResourceBindings: {
@@ -18,8 +18,8 @@ export const EditDescription = ({
   return (
     <TextResource
       handleIdChange={handleIdChange}
-      label={getLanguageFromKey('ux_editor.modal_properties_description', language)}
-      placeholder={getLanguageFromKey('ux_editor.modal_properties_description_add', language)}
+      label={t('ux_editor.modal_properties_description')}
+      placeholder={t('ux_editor.modal_properties_description_add')}
       textResourceId={component.textResourceBindings?.description}
     />
   );
