+++ frontend/packages/ux-editor/src/components/config/editModal/EditHeaderSize.tsx
@@ -2,7 +2,7 @@ import Select from 'react-select';
 import React from 'react';
 import { PropertyLabel, selectStyles } from '../../../utils/render';
 import type { IGenericEditComponent } from '../componentConfig';
-import { getLanguageFromKey } from 'app-shared/utils/language';
+import { useText } from '../../../hooks';
 
 enum HeaderSize {
   S = 'h4',
@@ -11,8 +11,8 @@ enum HeaderSize {
 }
 
 // Todo: This should be called "level" instead of "size"
-export const EditHeaderSize = ({ handleComponentChange, component, language }: IGenericEditComponent) => {
-  const t = (key: string) => getLanguageFromKey(key, language);
+export const EditHeaderSize = ({ handleComponentChange, component }: IGenericEditComponent) => {
+  const t = useText();
   const sizes = [
     { value: HeaderSize.S, label: t('ux_editor.modal_header_type_h4') },
     { value: HeaderSize.M, label: t('ux_editor.modal_header_type_h3') },
