+++ frontend/packages/ux-editor/src/containers/EditContainer.tsx
@@ -6,7 +6,6 @@ import '../styles/index.css';
 import { getComponentTitleByComponentType, getTextResource, truncate } from '../utils/language';
 import { componentIcons } from '../components';
 import { FormLayoutActions } from '../features/formDesigner/formLayout/formLayoutSlice';
-import { getLanguageFromKey } from 'app-shared/utils/language';
 import type { FormComponentType, IAppState, IFormComponent } from '../types/global';
 import classes from './EditContainer.module.css';
 import { Button, ButtonColor, ButtonVariant } from '@altinn/altinn-design-system';
@@ -15,6 +14,9 @@ import cn from 'classnames';
 import { ConnectDragSource } from 'react-dnd';
 import { DragHandle } from '../components/DragHandle';
 import { DEFAULT_LANGUAGE } from 'app-shared/constants';
+import { useText } from '../hooks';
+import { textSelector } from '../selectors/textSelectors';
+import { textResourcesByLanguageSelector } from '../selectors/textResourceSelectors';
 
 export interface IEditContainerProps {
   component: IFormComponent;
@@ -30,6 +32,7 @@ export interface IEditContainerProps {
 
 export function EditContainer(props: IEditContainerProps) {
   const dispatch = useDispatch();
+  const t = useText();
 
   const [component, setComponent] = useState<IFormComponent>({
     id: props.id,
@@ -46,9 +49,9 @@ export function EditContainer(props: IEditContainerProps) {
 
   const GetLayoutOrderSelector = makeGetLayoutOrderSelector();
   const activeList = useSelector((state: IAppState) => state.formDesigner.layout.activeList);
-  const language = useSelector((state: IAppState) => state.appData.languageState.language);
+  const language = useSelector(textSelector);
   const orderList = useSelector((state: IAppState) => GetLayoutOrderSelector(state));
-  const textResources = useSelector((state: IAppState) => state.appData.textResources.resources[DEFAULT_LANGUAGE]);
+  const textResources = useSelector(textResourcesByLanguageSelector(DEFAULT_LANGUAGE));
 
   const handleComponentUpdate = (updatedComponent: IFormComponent): void => {
     setComponent({ ...updatedComponent });
@@ -150,7 +153,7 @@ export function EditContainer(props: IEditContainerProps) {
                   80
                 )
                 : getComponentTitleByComponentType(component.type, language) ||
-                getLanguageFromKey('ux_editor.component_unknown', language)}
+                t('ux_editor.component_unknown')}
             </div>
           )}
         </div>
