+++ frontend/packages/ux-editor/src/utils/component.ts
@@ -1,19 +1,19 @@
-export function getTextResourceByAddressKey(key: AddressKeys, language: any): string {
+export function getTextResourceByAddressKey(key: AddressKeys, t: (key: string) => string): string {
   switch (key) {
     case AddressKeys.address: {
-      return language['ux_editor.modal_configure_address_component_address'];
+      return t('ux_editor.modal_configure_address_component_address');
     }
     case AddressKeys.zipCode: {
-      return language['ux_editor.modal_configure_address_component_zip_code'];
+      return t('ux_editor.modal_configure_address_component_zip_code');
     }
     case AddressKeys.houseNumber: {
-      return language['ux_editor.modal_configure_address_component_house_number'];
+      return t('ux_editor.modal_configure_address_component_house_number');
     }
     case AddressKeys.careOf: {
-      return language['ux_editor.modal_configure_address_component_care_of'];
+      return t('ux_editor.modal_configure_address_component_care_of');
     }
     case AddressKeys.postPlace: {
-      return language['ux_editor.modal_configure_address_component_post_place'];
+      return t('ux_editor.modal_configure_address_component_post_place');
     }
     default: {
       return '';
